add customer plan form

+++ src/main/java/com/akshata/entity/Plan.java
@@ -14,8 +14,8 @@ public class Plan {
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
 	private String plan_name;
-	private long pin_rate;
-	private long customer_amount;
+	private Double pin_rate;
+	private Double customer_amount;
 	private boolean status = true;
 	public long getId() {
 		return id;
@@ -29,16 +29,16 @@ public class Plan {
 	public void setPlan_name(String plan_name) {
 		this.plan_name = plan_name;
 	}
-	public long getPin_rate() {
+	public Double getPin_rate() {
 		return pin_rate;
 	}
-	public void setPin_rate(long pin_rate) {
+	public void setPin_rate(Double pin_rate) {
 		this.pin_rate = pin_rate;
 	}
-	public long getCustomer_amount() {
+	public Double getCustomer_amount() {
 		return customer_amount;
 	}
-	public void setCustomer_amount(long customer_amount) {
+	public void setCustomer_amount(Double customer_amount) {
 		this.customer_amount = customer_amount;
 	}
 	public boolean isStatus() {
